src/requirements.txt
click==7.1.2
coloredlogs==15.0
python-decouple==3.4
python-ldap==3.3.1